azurebatch/requirements.txt
# Azure Storage SDK for Python
azure-storage-blob==12.19.0

# Azure Identity for Managed Identity support
azure-identity==1.15.0

# Image processing library
Pillow==10.2.0
